refactor(core): split AssistantConfig/AgentConfig into types::agent submodule
Final step of the `core/src/types.rs` kitchen-sink split. Moves the
top-level `AssistantConfig` (root TOML schema) and `AgentConfig`
(active-agent context) plus `default_agent_id` into
`assistant_core::types::agent::*`, dropping them from the `lib.rs`
re-export.

`types.rs` now contains only:
- `pub(crate) fn default_true()` helper (shared with submodules via
  `super::default_true`)
- `pub mod {transcription, storage, observability, features, skills_mcp,
  channels, llm, conversation, agent}` declarations
- The remaining conversation type bodies (on this branch — moved by the
  parallel `types-split-conversation` PR #756)
- Test module imports updated to pull from the new submodules

Call sites updated across `crates/{core,tool-executor,runtime,web-ui,
interface-cli,interfaces,storage,mcp-client}` to import `AssistantConfig`
and `AgentConfig` via `assistant_core::types::agent::*`.

Tests:
- `cargo test -p assistant-core --lib` — 183 pass
- `cargo test -p assistant-runtime --lib` — 220 pass
- `cargo clippy --workspace -- -D warnings` — green

After this and the conversation PR merge, `types.rs` is purely a
namespace module — every type lives in its domain submodule.
