+++ crates/runtime/src/memory_indexer/mod.rs
@@ -10,7 +10,9 @@ use std::path::{Path, PathBuf};
 use std::sync::Arc;
 
 use anyhow::Result;
-use assistant_core::{AssistantConfig, LlmProvider, base_dir, resolve_dir, resolve_path};
+use assistant_core::{
+    LlmProvider, base_dir, resolve_dir, resolve_path, types::agent::AssistantConfig,
+};
 use assistant_storage::{MemoryChunkStore, StorageLayer};
 use sha2::{Digest, Sha256};
 use tracing::{debug, info, warn};
